LayoutTests/http/tests/site-isolation/paste-in-iframe.html

<!DOCTYPE html><!-- webkit-test-runner [ SiteIsolationEnabled=true ] -->
<script src="/js-test-resources/js-test.js"></script>
<script>
description("Tests that the iframe is able to receive paste events.");
jsTestIsAsync = true;

addEventListener("message", (event) => {
    shouldBe("event.data", "'Text to be copied and pasted'");
    finishJSTest();
});
</script>
<iframe src="http://localhost:8000/site-isolation/resources/copy-and-paste-text.html"></iframe>
